This is a poc for https://github.com/octokit/webhooks.js/pull/932

I think I will also take some parts of this PR to the beta. E.g. loglevel is not passed through every layer, resulting in logging everything, which can slow down significantly. 

I used npm link on the @octokit/webhooks.js branch I proposed to test it here. So the ci wont work.

Anyway here some naive benchmarks.

before:
```

┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max   │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼───────┤
│ Latency │ 1 ms │ 2 ms │ 6 ms  │ 7 ms │ 2.28 ms │ 1.45 ms │ 40 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴───────┘
┌───────────┬────────┬────────┬────────┬────────┬─────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%  │ Avg     │ Stdev  │ Min    │
├───────────┼────────┼────────┼────────┼────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 2167   │ 2167   │ 3813   │ 4009   │ 3607.91 │ 546.98 │ 2166   │
├───────────┼────────┼────────┼────────┼────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 321 kB │ 321 kB │ 565 kB │ 593 kB │ 534 kB  │ 81 kB  │ 321 kB │
└───────────┴────────┴────────┴────────┴────────┴─────────┴────────┴────────┘

Req/Bytes counts sampled once per second.
# of samples: 11

40k requests in 11.01s, 5.87 MB read
```

Because logging is not passed through the original performance is about 30 k/ 10 secs.

after:

```
Running 10s test @ http://127.0.0.1:3000/
10 connections


┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max   │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼───────┤
│ Latency │ 0 ms │ 0 ms │ 4 ms  │ 4 ms │ 0.57 ms │ 1.04 ms │ 31 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴───────┘
┌───────────┬────────┬────────┬─────────┬─────────┬────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%   │ Avg    │ Stdev  │ Min    │
├───────────┼────────┼────────┼─────────┼─────────┼────────┼────────┼────────┤
│ Req/Sec   │ 4727   │ 4727   │ 8047    │ 8187    │ 7675.2 │ 994.42 │ 4726   │
├───────────┼────────┼────────┼─────────┼─────────┼────────┼────────┼────────┤
│ Bytes/Sec │ 591 kB │ 591 kB │ 1.01 MB │ 1.02 MB │ 959 kB │ 124 kB │ 591 kB │
└───────────┴────────┴────────┴─────────┴─────────┴────────┴────────┴────────┘

Req/Bytes counts sampled once per second.
# of samples: 10

77k requests in 10.02s, 9.59 MB read
```